refactor: Rename user service imports for consistency and add useEditUser hook for managing user edits
