+++ frontend_web/react_frontend/src/modules/users/hooks/useCreateUser.js
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { createUser } from "../services/createUser";
+import { createUser } from "../services/createUserService";
 
 export function useCreateUser(formData) {
   const [form, setForm] = useState(formData);
