+++ frontend_web/react_frontend/src/modules/users/hooks/useEditUser.js
@@ -1,0 +1,8 @@
+export function useEditUser() {
+  function handleChange(e) {
+    setForm((prev) => ({
+      ...prev,
+      [e.target.name]: e.target.value,
+    }));
+  }
+}
