+++ frontend_web/react_frontend/src/modules/users/hooks/useUsers.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { getUsers } from "../services/getUsers";
+import { getUsers } from "../services/getUsersService";
 
 export function useUsers() {
   const [users, setUsers] = useState([]);
